+++ components/organisms/ContributorProfilePage/contributor-profile-page.tsx
@@ -99,8 +99,10 @@ const ContributorProfilePage = ({
         <SkeletonWrapper height={200} />
       ) : (
         <ContributorProfileHeader
+          isPremium={user ? user.role >= 50 : false}
           handleSignIn={signIn}
           username={user?.login}
+          isRecievingCollaborations={user?.receive_collaboration}
           user={loggedInUser}
           isFollowing={followError ? false : true}
           isConnected={iscConnected}